Split `core.models.gateways` in two
